Внесены правки от ревьюера
