+++ MovieQuiz/Services/NetworkClient.swift
@@ -1,5 +1,5 @@
 import Foundation
-/// Отвечает за загрузку данных по URL
+
 struct NetworkClient {
 
     private enum NetworkError: Error {
@@ -10,21 +10,18 @@ struct NetworkClient {
         let request = URLRequest(url: url)
     
         let task = URLSession.shared.dataTask(with: request) { data, response, error in
-            // Проверяем, пришла ли ошибка
             if let error = error {
                 handler(.failure(error))
                 return
             }
             
-            // Проверяем, что нам пришёл успешный код ответа
             if let response = response as? HTTPURLResponse,
                 response.statusCode < 200 || response.statusCode >= 300 {
                 handler(.failure(NetworkError.codeError))
                 return
             }
             
-            // Возвращаем данные
-            guard let data = data else { return }
+            guard let data else { return }
             handler(.success(data))
         }
